requirements.txt
numpy
pydicom
networkx
scikit-image
sklearn
pandas
python-igraph
